Add files via upload
The ATM System project is designed to simulate the functionality of an Automated Teller Machine. It allows users to perform various banking operations such as withdrawing cash, depositing funds, checking account balance, and receiving receipt . The project is implemented in Java programming language and utilizes object-oriented concepts for modular and efficient code organization.